gfaas-function-grpc-python3.9/requirements.txt
grpcio==1.54.0
grpcio-reflection==1.54.0
grpcio-tools==1.54.0
protobuf==4.22.3
reflection==0.1.1
